// src/types/winget.ts
export type InstallerType =
  | 'msi'
  | 'msix'
  | 'exe'
  | 'zip'
  | 'inno'
  | 'nullsoft'
  | 'wix'
  | 'burn'
  | 'portable'
  | 'unknown';

export type Architecture = 'x64' | 'x86' | 'arm64' | 'neutral';

export type Scope = 'user' | 'machine';

export interface InstallerSwitches {
  silent?: string;
  silentWithProgress?: string;
  interactive?: string;
  installLocation?: string;
  log?: string;
  upgrade?: string;
  custom?: string;
}

export interface Installer {
  architecture: Architecture;
  installerType: InstallerType;
  installerUrl: string;
  installerSha256?: string;
  scope?: Scope;
  installerSwitches?: InstallerSwitches;
  productCode?: string;
  upgradeBehavior?: string;
  installerLocale?: string;
  platform?: string[];
  minimumOsVersion?: string;
  nestedInstallerType?: InstallerType;
  nestedInstallerFiles?: { relativeFilePath: string }[];
}

export interface WingetManifest {
  packageIdentifier: string;
  packageName: string;
  packageVersion: string;
  publisher: string;
  publisherUrl?: string;
  publisherSupportUrl?: string;
  privacyUrl?: string;
  author?: string;
  moniker?: string;
  license?: string;
  licenseUrl?: string;
  copyright?: string;
  copyrightUrl?: string;
  shortDescription?: string;
  description?: string;
  tags?: string[];
  releaseNotes?: string;
  releaseNotesUrl?: string;
  installers: Installer[];
  defaultLocale?: string;
}